pyproject.toml
[project]
name = "jira-notion-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "numpy>=2.2.6",
    "openai>=1.107.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.4.0",
    "streamlit>=1.49.1",
]


